moved new classes to validation
